nada2

+++ demo/src/main/java/com/javeriana/taller_reactivo/demo/modelo/Nota.java
@@ -7,8 +7,6 @@ import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotNull;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @NoArgsConstructor
@@ -28,7 +26,7 @@ public class Nota {
     
     @DecimalMin("0.0")
     @DecimalMax("5.0")
-    private BigDecimal valor;
+    private double valor;
 
     @DecimalMin("0.0")
     @DecimalMax("100.0")
